chore: Client로부터의 Request Class 패키지 Application -> Presentation로 이동
