+++ src/main/java/project/reviewing/member/command/application/MemberService.java
@@ -3,9 +3,9 @@ package project.reviewing.member.command.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import project.reviewing.member.command.application.request.ReviewerRegistrationRequest;
-import project.reviewing.member.command.application.request.ReviewerUpdateRequest;
-import project.reviewing.member.command.application.request.MyInformationUpdateRequest;
+import project.reviewing.member.presentation.request.ReviewerRegistrationRequest;
+import project.reviewing.member.presentation.request.ReviewerUpdateRequest;
+import project.reviewing.member.presentation.request.MyInformationUpdateRequest;
 import project.reviewing.member.command.domain.Member;
 import project.reviewing.member.command.domain.MemberRepository;
 import project.reviewing.member.exception.MemberNotFoundException;
