+++ src/main/java/project/reviewing/member/presentation/MemberController.java
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import project.reviewing.auth.presentation.AuthenticatedMember;
 import project.reviewing.member.command.application.MemberService;
-import project.reviewing.member.command.application.request.MyInformationUpdateRequest;
-import project.reviewing.member.command.application.request.ReviewerRegistrationRequest;
-import project.reviewing.member.command.application.request.ReviewerUpdateRequest;
+import project.reviewing.member.presentation.request.MyInformationUpdateRequest;
+import project.reviewing.member.presentation.request.ReviewerRegistrationRequest;
+import project.reviewing.member.presentation.request.ReviewerUpdateRequest;
 import project.reviewing.member.query.application.MemberQueryService;
 import project.reviewing.member.query.application.response.MyReviewerInformationResponse;
 import project.reviewing.member.query.application.response.MyInformationResponse;
