+++ src/test/java/project/reviewing/integration/member/application/MemberServiceTest.java
@@ -11,9 +11,9 @@ import org.junit.jupiter.api.Test;
 import project.reviewing.common.exception.ErrorType;
 import project.reviewing.integration.IntegrationTest;
 import project.reviewing.member.command.application.MemberService;
-import project.reviewing.member.command.application.request.MyInformationUpdateRequest;
-import project.reviewing.member.command.application.request.ReviewerRegistrationRequest;
-import project.reviewing.member.command.application.request.ReviewerUpdateRequest;
+import project.reviewing.member.presentation.request.MyInformationUpdateRequest;
+import project.reviewing.member.presentation.request.ReviewerRegistrationRequest;
+import project.reviewing.member.presentation.request.ReviewerUpdateRequest;
 import project.reviewing.member.command.domain.Member;
 import project.reviewing.member.command.domain.Reviewer;
 import project.reviewing.member.exception.MemberNotFoundException;
